% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_desktop_launcher.R
\name{create_desktop_launcher}
\alias{create_desktop_launcher}
\title{Create a desktop launcher directory and .bat via shiny.exe}
\usage{
create_desktop_launcher(path = "QCcorrection_app", app_name = "QCcorrection")
}
\description{
Create a desktop launcher directory and .bat via shiny.exe
}
