% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_launcher_bat.R
\name{make_launcher_bat}
\alias{make_launcher_bat}
\title{Create a desktop launcher directory and .bat via shiny.exe}
\usage{
make_launcher_bat(
  dir = system.file("launcher", package = "QCcorrection"),
  bat_name = "QCcorrection.bat",
  app_file = "app.R"
)
}
\description{
Create a desktop launcher directory and .bat via shiny.exe
}
