% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCcorrection-package.R
\docType{package}
\name{QCcorrection-package}
\alias{QCcorrection}
\alias{QCcorrection-package}
\title{QCcorrection: Shiny app for QC correction of metabolomics data}
\description{
Shiny app for QC correction of metabolomics data
}
\author{
\strong{Maintainer}: Breanna Guppy \email{breanna-guppy@uiowa.edu}

}
\keyword{internal}
